remove duplicate class and move test case to amoro-format-iceberg
